Use newer PoolMath API
The newer PoolMath API helps minimize the impact of API calls on the backend server.  It also helps with reliability as it uses the `poolId` which is more stable than the share code.

Also increased the interval to 8 minutes by default since the service caches the data for 10 minutes anyway.